Enable `ruff`

+++ tests/plotting/test_utils.py
@@ -32,9 +32,8 @@ class TestMoscotPl:
     @pytest.mark.parametrize("save", [None, "tests/data/test_plot.png"])
     @pytest.mark.parametrize("return_fig", [True, False])
     def test_cell_transition(self, adata_pl_cell_transition: AnnData, return_fig: bool, save: Optional[str]):
-        if save:
-            if os.path.exists(save):
-                os.remove(save)
+        if save and os.path.exists(save):
+            os.remove(save)
         fig = msc.plotting.cell_transition(adata_pl_cell_transition, return_fig=return_fig, save=save)
         if return_fig:
             assert fig is not None
@@ -50,9 +49,8 @@ class TestMoscotPl:
     def test_push(
         self, adata_pl_push: AnnData, time_points: Optional[List[int]], return_fig: bool, save: Optional[str]
     ):
-        if save:
-            if os.path.exists(save):
-                os.remove(save)
+        if save and os.path.exists(save):
+            os.remove(save)
 
         fig = msc.plotting.push(adata_pl_push, time_points=time_points, return_fig=return_fig, save=save)
 
@@ -70,9 +68,8 @@ class TestMoscotPl:
     def test_pull(
         self, adata_pl_pull: AnnData, time_points: Optional[List[int]], return_fig: bool, save: Optional[str]
     ):
-        if save:
-            if os.path.exists(save):
-                os.remove(save)
+        if save and os.path.exists(save):
+            os.remove(save)
         fig = msc.plotting.pull(adata_pl_pull, time_points=time_points, return_fig=return_fig, save=save)
         if return_fig:
             assert fig is not None
@@ -86,9 +83,8 @@ class TestMoscotPl:
     @pytest.mark.parametrize("return_fig", [True, False])
     @pytest.mark.parametrize("interpolate_color", [True, False])
     def test_sankey(self, adata_pl_sankey: AnnData, return_fig: bool, save: Optional[str], interpolate_color: bool):
-        if save:
-            if os.path.exists(save):
-                os.remove(save)
+        if save and os.path.exists(save):
+            os.remove(save)
         fig = msc.plotting.sankey(
             adata_pl_sankey,
             return_fig=return_fig,
